[refactor] 코드 및 패키지, 파일 정리
