+++ app/src/main/java/org/sopt/soptseminar/data/repositories/DefaultUserAuthRepository.kt
@@ -1,9 +1,9 @@
 package org.sopt.soptseminar.data.repositories
 
-import org.sopt.soptseminar.data.service.SoptService
+import org.sopt.soptseminar.data.services.SoptService
 import org.sopt.soptseminar.data.models.sign.RequestSignIn
-import org.sopt.soptseminar.domain.UserAuthRepository
-import org.sopt.soptseminar.models.UserInfo
+import org.sopt.soptseminar.domain.repositories.UserAuthRepository
+import org.sopt.soptseminar.domain.models.UserInfo
 import org.sopt.soptseminar.modules.datastore.UserPreferenceRepository
 import javax.inject.Inject
