--- app/src/main/java/org/sopt/soptseminar/data/service/GithubServiceCreator.kt
@@ -1,14 +1,0 @@
-package org.sopt.soptseminar.data.service
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object GithubServiceCreator {
-    private const val BASE_URL = "https://api.github.com/"
-
-    private val retrofit: Retrofit = Retrofit.Builder().baseUrl(BASE_URL).addConverterFactory(
-        GsonConverterFactory.create()
-    ).build()
-
-    val githubService: GithubService = retrofit.create(GithubService::class.java)
-}
