+++ app/src/main/java/org/sopt/soptseminar/data/services/GithubService.kt
@@ -1,4 +1,4 @@
-package org.sopt.soptseminar.data.service
+package org.sopt.soptseminar.data.services
 
 import org.sopt.soptseminar.data.models.github.ResponseFollower
 import org.sopt.soptseminar.data.models.github.ResponseRepository
