+++ app/src/main/java/org/sopt/soptseminar/domain/di/ServiceBinder.kt
@@ -4,24 +4,31 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import org.sopt.soptseminar.data.service.GithubService
-import org.sopt.soptseminar.data.service.GithubServiceCreator
-import org.sopt.soptseminar.data.service.SoptService
-import org.sopt.soptseminar.data.service.SoptServiceCreator
+import org.sopt.soptseminar.data.services.GithubService
+import org.sopt.soptseminar.data.services.SoptService
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-object ServiceBinder {
+object RetrofitBinder {
+    private const val GITHUB_BASE_URL = "https://api.github.com/"
+    private const val SIGN_BASE_URL = "http://13.124.62.236/"
+
     @Singleton
     @Provides
     fun bindSoptService(): SoptService {
-        return SoptServiceCreator.soptService
+        return Retrofit.Builder().baseUrl(SIGN_BASE_URL).addConverterFactory(
+            GsonConverterFactory.create()
+        ).build().create(SoptService::class.java)
     }
 
     @Singleton
     @Provides
     fun bindGithubService(): GithubService {
-        return GithubServiceCreator.githubService
+        return Retrofit.Builder().baseUrl(GITHUB_BASE_URL).addConverterFactory(
+            GsonConverterFactory.create()
+        ).build().create(GithubService::class.java)
     }
 }
