+++ app/src/main/java/org/sopt/soptseminar/domain/models/github/FollowerInfo.kt
@@ -1,12 +1,13 @@
-package org.sopt.soptseminar.models
+package org.sopt.soptseminar.domain.models.github
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class RepositoryInfo(
+data class FollowerInfo(
     val name: String,
-    val description: String? = null,
+    val profile: String,
     val url: String,
+    val description: String? = null,
     val id: Int
 ) : Parcelable
