+++ app/src/main/java/org/sopt/soptseminar/domain/models/github/RepositoryInfo.kt
@@ -1,13 +1,12 @@
-package org.sopt.soptseminar.models
+package org.sopt.soptseminar.domain.models.github
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class FollowerInfo(
+data class RepositoryInfo(
     val name: String,
-    val profile: String,
-    val url: String,
     val description: String? = null,
+    val url: String,
     val id: Int
 ) : Parcelable
