+++ app/src/main/java/org/sopt/soptseminar/domain/models/types/SoptPartType.kt
@@ -1,4 +1,4 @@
-package org.sopt.soptseminar.models.types
+package org.sopt.soptseminar.domain.models.types
 
 import org.sopt.soptseminar.R
