--- app/src/main/java/org/sopt/soptseminar/models/SignInfo.kt
@@ -1,10 +1,0 @@
-package org.sopt.soptseminar.models
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class SignInfo(
-    val id: String,
-    val password: String,
-) : Parcelable
