+++ app/src/main/java/org/sopt/soptseminar/modules/datastore/UserPreferenceRepository.kt
@@ -1,7 +1,7 @@
 package org.sopt.soptseminar.modules.datastore
 
 import kotlinx.coroutines.flow.Flow
-import org.sopt.soptseminar.models.UserInfo
+import org.sopt.soptseminar.domain.models.UserInfo
 
 interface UserPreferenceRepository {
     suspend fun getUsersPreference(): Flow<UserInfo?>
