+++ app/src/main/java/org/sopt/soptseminar/presentation/github/adapters/GithubAdapter.kt
@@ -3,7 +3,7 @@ package org.sopt.soptseminar.presentation.github.adapters
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import org.sopt.soptseminar.models.types.GithubDetailViewType
+import org.sopt.soptseminar.presentation.types.GithubDetailViewType
 import org.sopt.soptseminar.presentation.github.screens.FollowerFragment
 import org.sopt.soptseminar.presentation.github.screens.RepositoryFragment
