+++ app/src/main/java/org/sopt/soptseminar/presentation/github/viewmodels/GithubViewModel.kt
@@ -8,10 +8,10 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
-import org.sopt.soptseminar.domain.GithubProfileRepository
-import org.sopt.soptseminar.models.FollowerInfo
-import org.sopt.soptseminar.models.RepositoryInfo
-import org.sopt.soptseminar.models.UserInfo
+import org.sopt.soptseminar.domain.repositories.GithubProfileRepository
+import org.sopt.soptseminar.domain.models.github.FollowerInfo
+import org.sopt.soptseminar.domain.models.github.RepositoryInfo
+import org.sopt.soptseminar.domain.models.UserInfo
 import org.sopt.soptseminar.modules.datastore.UserPreferenceRepository
 import javax.inject.Inject
