+++ app/src/main/res/layout/item_repository.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="repository"
-            type="org.sopt.soptseminar.models.RepositoryInfo" />
+            type="org.sopt.soptseminar.domain.models.github.RepositoryInfo" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
